LayoutTests/imported/w3c/web-platform-tests/css/css-anchor-position/position-try-custom-property-expected.txt


FAIL .target 1 assert_equals:
<div class="target fallback1" data-offset-x="150" data-offset-y="25"></div>
offsetLeft expected 150 but got 5
FAIL .target 2 assert_equals:
<div class="target fallback2" data-offset-x="150" data-offset-y="25"></div>
offsetLeft expected 150 but got 5
